backend/config.py
import os
from dotenv import load_dotenv
from datetime import timedelta




load_dotenv()

JWT_ACCESS_TOKEN_EXPIRES = timedelta(hours=1)
class Config:
    SECRET_KEY = os.getenv("SECRET_KEY")
    JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY")
    MONGO_URI = os.getenv("MONGO_URI")
